+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.31.2](https://github.com/sidthesloth92/db-astro-suite/compare/v1.31.1...v1.31.2) (2026-06-28)
+
+
+### Bug Fixes
+
+* **starwizz:** hold loading spinner until default image loads; remove fake HUD telemetry ([#94](https://github.com/sidthesloth92/db-astro-suite/issues/94)) ([e243a72](https://github.com/sidthesloth92/db-astro-suite/commit/e243a72ec02bf2b66d8b1733e535df304fb2efa7))
+
+## [1.31.1](https://github.com/sidthesloth92/db-astro-suite/compare/v1.31.0...v1.31.1) (2026-06-21)
+
+
+### Bug Fixes
+
+* **hub:** play landing intro once per session to stop Back-nav replay ([#92](https://github.com/sidthesloth92/db-astro-suite/issues/92)) ([5d8539d](https://github.com/sidthesloth92/db-astro-suite/commit/5d8539d41ffa568cddd069b21a84024b50271d0f))
+
 ## [1.31.0](https://github.com/sidthesloth92/db-astro-suite/compare/v1.30.0...v1.31.0) (2026-06-20)